section 2 exo 5

+++ python_fundamentals-master/python_fundamentals-master/02_basic_datatypes/1_numbers/02_04_temp.py
@@ -11,5 +11,5 @@ Output should read like - "81.32 degrees fahrenheit = 27.4 degrees celsius"
 
 '''
 Fahrenheit = (input("Enter your Fahrenheit degree there: "))
-celsius = (int(Fahrenheit) - 32) * (5 / 9)
+celsius = (float(Fahrenheit) - 32) * (5 / 9)
 print(Fahrenheit, "degrees fahrenheit =", round(celsius, 1), "degrees celsius")
